% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlvizstats-package.R
\docType{package}
\name{mlvizstats-package}
\alias{mlvizstats}
\alias{mlvizstats-package}
\title{mlvizstats: What the Package Does (Title Case)}
\description{
Understanding machine learning using visualizations.
}
\keyword{internal}
